packages/codeflicker-agent/pyproject.toml
[project]
name = "codeflicker-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-a2a>=0.5.0",
    "openai>=1.30.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codeflicker_agent"]
